BaseModelとその依存関係を削除
