+++ ami/models/__init__.py
@@ -1,2 +1,2 @@
 """This file contains import statements for convenience."""
-from . import base_model, components, utils
+from . import components, utils
