+++ tests/conftest.py
@@ -5,29 +5,14 @@ import torch
 from ami.data.step_data import DataKeys, StepData
 from ami.data.utils import DataCollectorsDict, DataUsersDict
 from ami.models.model_wrapper import ModelWrapper
-from ami.models.utils import InferencesDict, ModelsDict, ModelWrappersDict
-from tests.helpers import DataBufferImpl, ModelImpl, ModelMultiplyP, get_gpu_device
+from ami.models.utils import ModelWrappersDict
+from tests.helpers import DataBufferImpl, ModelMultiplyP, get_gpu_device
 
 
 @pytest.fixture
 def gpu_device() -> torch.device | None:
     """Fixture for retrieving the available gpu device."""
     return get_gpu_device()
-
-
-@pytest.fixture
-def models_dict(gpu_device: torch.device | None) -> ModelsDict:
-    if gpu_device is not None:
-        device = gpu_device
-    else:
-        device = "cpu"
-
-    d = ModelsDict(
-        {"model1": ModelImpl("cpu", True), "model2": ModelImpl("cpu", False), "model3": ModelImpl(device, True)}
-    )
-
-    d.send_to_default_device()
-    return d
 
 
 @pytest.fixture
